requirements.txt
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
autocommand==2.2.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.10
inflect==7.3.1
magic-filter==1.0.12
more-itertools==10.8.0
multidict==6.6.4
oauthlib==3.3.1
packaging==24.2
platformdirs==4.2.2
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
qrcode==8.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
tomli==2.0.1
typeguard==4.3.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
zipp==3.19.2
Pillow==11.3.0
